move SetAsOptimized out of core
